# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: Laravel
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://laravel-api.test'
paths:
    /api/categories:
        get:
            summary: 'Get Categories'
            description: 'List all the categories.'
            parameters:
                -
                    in: query
                    name: page
                    description: 'Which page to show.'
                    example: 12
                    required: false
                    schema:
                        type: integer
                        description: 'Which page to show.'
                        example: 12
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: [{ id: 1, name: 'Category 1', created_at: '2021-02-25T20:17:16.000000Z' }, { id: 2, name: 'Category 2', created_at: '2021-02-25T20:17:16.000000Z' }, { id: 3, name: 'Category 3', created_at: '2021-02-25T20:17:16.000000Z' }, { id: 4, name: 'Category 4', created_at: '2021-02-25T20:17:16.000000Z' }]
                                properties:
                                    data: { type: array, example: [{ id: 1, name: 'Category 1', created_at: '2021-02-25T20:17:16.000000Z' }, { id: 2, name: 'Category 2', created_at: '2021-02-25T20:17:16.000000Z' }, { id: 3, name: 'Category 3', created_at: '2021-02-25T20:17:16.000000Z' }, { id: 4, name: 'Category 4', created_at: '2021-02-25T20:17:16.000000Z' }], items: { type: object } }
                404:
                    description: '404'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    message: 'Not found'
                                properties:
                                    message: { type: string, example: 'Not found' }
            tags:
                - Categories
            security: []
        post:
            summary: 'POST Categories'
            description: 'Create new category.'
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Categories
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'Name of the category.'
                                    example: '"Clothing"'
                                photo:
                                    type: string
                                    format: binary
                                    description: 'The value must be an image.'
                            required:
                                - name
            security: []
...